Add create alert button and begin to scaffold dialog
