+++ src/modules/client/components/clientAlerts/ClientAlertCard.tsx
@@ -1,7 +1,4 @@
 import { Alert, AlertTitle, Box, Typography } from '@mui/material';
-import { Stack } from '@mui/system';
-import Loading from '@/components/elements/Loading';
-import TitleCard from '@/components/elements/TitleCard';
 import theme from '@/config/theme';
 import {
   clientNameAllParts,
@@ -13,7 +10,7 @@ import {
   HouseholdClientFieldsClientFragment,
 } from '@/types/gqlTypes';
 
-interface ClientAlertCardProps {
+export interface ClientAlertCardProps {
   alert: ClientAlertFieldsFragment;
   client: ClientFieldsFragment | HouseholdClientFieldsClientFragment;
   shouldShowClientName?: boolean;
@@ -90,71 +87,4 @@ const ClientAlertCard: React.FC<ClientAlertCardProps> = ({
   );
 };
 
-export enum AlertContext {
-  Client = 'Client',
-  Household = 'Household',
-}
-
-interface ClientAlertsProps {
-  clients: ClientFieldsFragment[] | HouseholdClientFieldsClientFragment[];
-  alertContext: AlertContext;
-  loading?: boolean;
-}
-
-const ClientAlerts: React.FC<ClientAlertsProps> = ({
-  clients,
-  alertContext = AlertContext.Client,
-  loading = true,
-}) => {
-  const clientAlerts: ClientAlertCardProps[] = [];
-  if (!clients) return;
-  clients.forEach((c) => {
-    if (c.access.canViewClientAlerts) {
-      c.alerts.forEach((a) => {
-        clientAlerts.push({
-          alert: a,
-          client: c,
-          shouldShowClientName: alertContext === AlertContext.Household,
-        });
-      });
-    }
-  });
-
-  const title = `${alertContext} Alerts (${clientAlerts.length})`;
-  if (loading && clients.length === 0) return <Loading />;
-
-  return (
-    <TitleCard
-      title={title}
-      headerVariant='border'
-      headerTypographyVariant='body1'
-    >
-      <Box sx={{ m: 2 }}>
-        {clientAlerts.length === 0 && (
-          <Box
-            sx={{
-              p: 2,
-              backgroundColor: '#fafafa',
-              color: '#00000060',
-              borderRadius: 1,
-            }}
-          >
-            {alertContext} has no alerts at this time
-          </Box>
-        )}
-        <Stack gap={2}>
-          {clientAlerts.map((ca) => (
-            <ClientAlertCard
-              key={ca.alert.id}
-              alert={ca.alert}
-              client={ca.client}
-              shouldShowClientName={ca.shouldShowClientName}
-            />
-          ))}
-        </Stack>
-      </Box>
-    </TitleCard>
-  );
-};
-
-export default ClientAlerts;
+export default ClientAlertCard;
